changed to Naive Bayes classifier
KNN was not working so I added some evaluative code to the training
method. After doing that, I played around with different classifiers and
found that Naive Bayes performs well with the data we have. If anyone
actually reads this, note that this is not the code or classifier we are
currently using in the project.